+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 * nostr: cleanup error enums ([Yuki Kishimoto])
 * nostr: use `SECP256K1` global context from `secp256k1` crate ([Yuki Kishimoto])
 * nostr: manually implement `PartialEq`, `Eq`, `PartialOrd`, `Ord` and `Hash` for `TagKind` ([Yuki Kishimoto])
+* nostr: derive `PartialEq` and `Eq` for event builder error ([Yuki Kishimoto])
 * pool: update `Error::WebSocket` variant inner type ([Yuki Kishimoto])
 * lmdb: use `EventBorrow` instead of `DatabaseEvent` ([Yuki Kishimoto])
 * ndb: refactor note-to-event conversion ([Yuki Kishimoto])